Add support to StaticFile so that it works with non-textual files.
Previously binary files were being interpreted as UTF8 encoded bytestreams
this isn't an ideal for pushing out static files such as images.

I've modified the StaticFile type so that it explicitly declares the encoding
as 'binary' and updated the mechanism used to push back to the client to
also be 'binary aware'.  I'm not happy with my approach as I've violated the
encapsulation of the logic that was previously in the request.respond method, but
without further direction on how that object is moving this was the easiest thing to do
in order to keep moving with my plan to have a full website ;)