Do not lie about what res.contentType() accepts
Passing it a literal content type string will not always work. For
`application/json` it works only by pure chance. It is because `json` is also
coincidentally the file extension. But if you tried `application/javascript`
it would set the content type header to `application/octet-stream`, because
`javascript` is not a known file extension.