ETag support in HEAD requests
I'm assuming that ETags should also be sent in the HEAD requests. And it would also be quite helpful in the job I have in hands at the moment :) 

According to the RFC (http://www.w3.org/Protocols/rfc2616/rfc2616-sec9.html#sec9.4) "The HEAD method is identical to GET except that the server MUST NOT return a message-body in the response". It even mentions "Content-Length, Content-MD5, ETag or Last-Modified" headers in HEAD requests.