Fix: 'trust proxy' setting not inherited by mounted app.
The issue was introduced in 566720be where app.defaultConfiguration was updated
to set 'trust proxy' unconditionally, which means that the default value of false
would always shadow the value that should be inherited from the parent app.

This fix avoids setting a default value and instead handles the absence of
a 'trust proxy fn' setting whenever it's used.

Fixes #2550.