tests: support IANA compliant header messages
This commit allows the tests to pass when the node HTTP server uses IANA
compliant header messages rather than the messages the server used to
use. This prevents future test breakage caused by
https://github.com/nodejs/io.js/pull/1470, a pull request tracked for
node 3.0.0 semver-major release. See:
https://github.com/nodejs/io.js/commit/235036e7faa537469c3600850bdd533c095c392a

The only pertinent status code change is the 302 code, which used to be
"Moved Temporarily" but is now "Found", according to the document here:
http://www.iana.org/assignments/http-status-codes/http-status-codes.xhtml