Security improvement: don't reveal powered-by
The first thing recommended when setting up an Express instance is to
secure it by removing its "X-Powered-By" header.

https://www.npmjs.com/package/helmet
https://blog.risingstack.com/node-js-security-checklist/
https://strongloop.com/strongblog/best-practices-for-express-in-production-part-one-security/

So let's save cycles by not setting this header in the first place.

This header is also totally useless. If some people are needing it
somewhere for a very specific usecase they can add a little custom
middleware that returns the "X-Powered-By" header again.