Add AUTHORS file and update script
tools/update_authors.js reads the git history (using the git command
line client) and processes the output to remove duplicate entries
(keyed from the user's email address).

The resulting list of authors is written to a file AUTHORS in the
current working directory. Author entries are in order of first
commit but use the most recent name for the given email address.