Update version to 4.13.4
NOTE: Version 4.14.4 was originally released to NPM based on
the expressjs/express repository (a, now deleted, github fork
of this repository which appears to have been based on
commit 4fb3653b6ce9bd9080ceadb267f2cffc32413444).

The following commits (ancestors to this one) recreate the
changes to History.md and package.json:
f9b4654050aee6a51b41dec91e4f6e872d5d438e Order lists alphabetically
99573b701eb97c59a8c03e27e12a9ce9cd4c7566 build: mocha@2.3.4
04ba299984ee5c012d93ee0e6d1438bd0da84c22 build: express-session@~1.13.0
65be7f32993b91a827d7943c591a021e9807e080 build: cookie-parser@~1.4.1
25e11dbba60587b0c5074d14ecb9aa0832f2f882 build: body-parser@~1.14.2
65d00eb1d296f1462fb2451e6d1bcfaa0d234fdc build: istanbul@0.4.2
0ab598454a48159035ae08935016c30bdb5303ee deps: serve-static@~1.10.2
69b0ff54d5e741366df992d434141c7f31c9c6dc deps: send@0.13.1
6255bec69ad9c0a058bf604d3b4a3856cb559c69 deps: range-parser@~1.0.3
66140b403ae7b0c452a34b9d4f9443041e5ee196 deps: proxy-addr@~1.0.10
31e73fc1cd37d3a7f4bdcddca805e848b0ec559d deps: parseurl@~1.3.1
40744efb32cddeb1d6ae4ba674a3adb1ed07ab05 deps: methods@~1.1.2
4e333fbc37099d7c53356c88eaf2dabc44fdaca8 deps: merge-descriptors@1.0.1
914fb8fbbd42d209a6ef557e8d9cc2d63b1fb369 deps: finalhandler@0.4.1
1827559e4453f85ed9faea9914e6b5e5a7ab9929 deps: escape-html@~1.0.3
fcefa48bb0890acfd2f1868411bccdb51d96240a deps: depd@~1.1.0
47dd5d6aed9cf7a5c3638ef84c04b2c3a590e479 deps: content-disposition@0.5.1

There were also changes to Readme.md, but these have not been
included in this series of commits as they change a number of
links to point to the now deleted repository (see #2861 and #2862).

This series of commits should move the repository forward to
a point where we can be happy most of the changes from 4.13.4
are included which is a prerequisite to releasing the next
release.

I hope this helps with any potential future confusion regarding
the differences between the npm package for 4.13.4 and what is
in this repository (for example, the differences to Readme.md,
the missing/different commit hashes, etc).