Use mime-types instead of send.mime
The mime-types module has a few small differences compared
to mime:

* mime-types API for getting charsets is slightly different
  to mime

* mime-types uses mime-db which has a comprehensive list of
  charsets for mime types, whereas mime uses a simple rule:
  UTF-8 for any type starting 'text/' and a caller-defined
  fallback for anything else

* mime-types will return false if it cannot find a particular
  mime type in mime-db, whereas mime will default to
  'application/octet-stream' if it does't find the mime type
  in its types.json file

This commit makes the necessary changes to address these
differences.

The mime module to use is now populated by
require('mime-types') rather than reaching into the send
module with require('send').mime.

One of the tests has been updated as a result of the charset
lookup differences in mime-types, since mime-types gives
UTF-8 as the charset for 'application/javascript' whereas
mime returned the caller fallback (undefined in this case.)
From what I can tell both the resulting values are valid in
the Content-Type header (either 'Content-Type:
application/javascript; charset=utf=8' or 'Content-Type:
application/javascript') so I have updated the testcase to
test for the former instead of the latter.