Proper usage of supertest expect in tracing tests
Supertest lib provides a method called expect to test the result of a request.
This method requires a callback that takes an error object as first parametr.
If this error object is not inspected, the validation performed by expect are
not checked. Tracing tests didn't do that checking.

This commit add checking to trace tests to ensure that no error is returned
when expect method is called.