fix several examples
**1. examples/markdown**

http://localhost:3000/fail
Error: Failed to lookup view "missing" in views directory

**2. examples/params**

http://localhost:3000/users/0-2
users tj, tobi     <-- lost loki

Because user/0 is tj, that’s means the 0 is the data, not index of
array.
So, users/0-2 must be include user/2.

**3. examples/route-separation**

http://localhost:3000
Click link "posts", jumped to the "users" page.

**4. examples/view-constructor**

http://localhost:3000/
ReferenceError: https is not defined

**5. examples/online, examples/search, and examples/session**

// first (if you have not done it yes):
// $ npm install redis
// $ brew install redis   <-- lost this line
// $ redis-server

If the user has installed redis, then there is no problem reinstalling
redis. But if here does not prompt the user to install redis, this
example will run failure, the user will be confused.