Fix cookie not clearing if maxAge is set Since maxAge takes precedence over expires, if maxAge was set the cookie would not expire.
Changed merge parameters' orders so that the cookie's options are overrided by {maxAge:0}. Note "expires: new Date(1)" has been removed as it is now obsolete since when 'maxAge' is set, 'expires' will also be set.
Also updated logic relating to setting 'expires' property when 'maxAge' is set. If 'maxAge' is <=0 it will set 'expires' to a past date. Otherwise, set it to the current Date + 'maxAge'.
Added appropriate tests checking the 'expires' and 'maxAge' properties are overriden by clearCookie(). Also updated existing tests so that they now expect a Max-Age=0; that is always set when a cookie is cleared now.