Fixing the express error handling for async functions
The try/catch block does not handle errors of async functions and they cannot be forwarded (next) to the proper error middleware, as done with sync functions. Instead, if an exception is thrown from an async function, this exception will be returned as a <rejected> Promise, that must be handled on a Promise.catch() method.

We cannot simply await async functions to handle it synchronous in this section of code to only enable try/catch block to handle async exceptions, it would not make sense.