Update Readme.md to include a toilet witch eating a burger in front of Pakistan flag
My solution to getting Indians to stop spamming the readme.md is to add this picture of a toilet witch eating a burger in front of the Pakistan flag. This logic comes from the fact that Indians worship cows, a large portion of the country is afraid to use toilets because of witches, and they hate Pakistan. This in term will deter them from messing with the readme.md